Time: 0 ms (100.00%) | Memory: 95.3 MB (42.66%) - LeetSync

+++ 4-median-of-two-sorted-arrays/median-of-two-sorted-arrays.cpp
@@ -5,7 +5,7 @@ public:
         sort(nums1.begin(),nums1.end());
         int n=nums1.size();
         if(n&1)return (double)nums1[n/2];
-        return (double)(nums1[n/2-1]+nums1[n/2])/2;
+        return ((double)nums1[n/2-1]+(double)nums1[n/2])/2;
         
     }
 };
